+++ cmd/warpbox/main.go
@@ -125,6 +125,8 @@ func main() {
 		throttleQueue,
 		time.Duration(cfg.Sync.IntervalMinutes)*time.Minute,
 		cfg.Sync.Limit,
+		*cfg.Sync.RetryAttempts,
+		time.Duration(*cfg.Sync.RetryBackoff)*time.Second,
 	)
 
 	// Set library change hooks.